Move schedule-controls
